requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.3
entrypoints==0.3
funcy==1.16
future==0.18.2
gensim==4.0.1
idna==2.10
install==1.3.4
ipykernel==5.5.5
ipython==7.10.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
kiwisolver==1.3.1
MarkupSafe==2.0.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
mecab-python3==1.0.3
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numexpr==2.7.3
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
pyLDAvis==3.3.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.1.0
RandomFileTree==1.2.0
requests==2.25.1
scipy==1.6.3
Send2Trash==1.5.0
six==1.16.0
smart-open==5.0.0
sniffio==1.2.0
sqlparse==0.4.1
terminado==0.9.5
testpath==0.5.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
